# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Next Swagger API Example",
    "version": "1.0"
  },
  "paths": {
...
    "/api/trending/scrapDB": {
      "get": {
        "summary": "Retrieve trending tags from the database for a specific provider and country.",
        "parameters": [
          {
            "in": "query",
            "name": "provider",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The provider of the trending tags."
          },
          {
            "in": "query",
            "name": "country",
            "schema": {
              "type": "string"
            },
            "description": "The country to get trending tags for. If not provided, it will return tags for all countries."
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response returns the trending tags for the specified provider and country."
          },
          "400": {
            "description": "Invalid query parameters."
          }
        }
      },
      "post": {
        "summary": "Delete all trending tags for a specific provider and insert new ones.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "provider": {
                    "type": "string",
                    "description": "The provider of the trending tags."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A successful response returns the inserted trending tags."
          }
        }
      }
    },
...
  },
...
}